requirements.txt
google-generativeai
speechrecognition
pyaudio
piper-tts
face-recognition
opencv-python
numpy
pyvrm
aiohttp
Pillow